webapprt/test/content/webapprt_sample.html

<!DOCTYPE HTML>

<!--
  This is a sample WebappRT content mochitest.  Since its name is prefixed with
  webapprt_, this file is picked up by the Mochitest harness.  It's just a plain
  mochitest that runs in the app browser within an app window.
-->

<html>
  <head>
    <meta charset="utf-8">
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>
  <body>
    <p id="display">
      This is the sample WebappRT content mochitest.
    </p>
    <div id="content" style="display: none"></div>
    <pre id="test">
      <script>

SimpleTest.waitForExplicitFinish();

ok(true, "true is true!");

var self = navigator.mozApps.getSelf();
self.onsuccess = function () {
  ok(true, "onsuccess should be called");
  ok(self.result, "result should be nonnull");
  ok(self.result.manifest, "manifest should be nonnull");
  is(self.result.manifest.name, "WebappRT Mochitest Webapp",
     "manifest.name");
  SimpleTest.finish();
};
self.onerror = function () {
  ok(false, "onerror should not be called");
  SimpleTest.finish();
};

      </script>
    </pre>
  </body>
</html>
